Move KitsuneP2pConfig into kitsune_p2p_types
